+++ partiql-value/src/lib.rs
@@ -52,7 +52,7 @@ pub enum Value {
     List(Box<List>),
     Bag(Box<Bag>),
     Tuple(Box<Tuple>),
-    // TODO: add other supported PartiQL values -- timestamp, date, time, sexp
+    // TODO: add other supported PartiQL values -- sexp
 }
 
 impl ops::Add for &Value {